third_party/blink/web_tests/http/tests/websocket/iframe-sandbox-expected.txt

WebSocket test that Origin is 'null' in a sandboxed iframe.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS wsOrigin is "null"
PASS successfullyParsed is true

TEST COMPLETE
